Refactor lib to src Remove examples folders Create dist folder Remove empty error on package.json Change build
