+++ src/steemconnect.js
